+++ tests/Feature/ArtisanTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Hyvor\LaravelPlaywright\Tests\Feature;
+namespace Advalis\LaravelPlaywright\Tests\Feature;
 
-use Hyvor\LaravelPlaywright\Tests\TestCase;
+use Advalis\LaravelPlaywright\Tests\TestCase;
 
 class ArtisanTest extends TestCase
 {